requirements.txt
# 核心依赖
arxiv==2.1.0                    # arXiv API 客户端
pyyaml==6.0.1                   # YAML 配置文件解析
python-dotenv==1.0.0            # 环境变量管理

# HTTP 客户端 (需要与 openai 兼容)
httpx==0.27.0                   # HTTP 客户端 (固定版本以兼容 openai 1.12.0)

# LLM 相关
openai==1.12.0                  # OpenAI API
anthropic==0.18.1               # Anthropic Claude API (可选)rxiv==2.1.0                    # arXiv API 客户端
pyyaml==6.0.1                   # YAML 配置文件解析
python-dotenv==1.0.0            # 环境变量管理

# LLM 相关
openai==1.12.0                  # OpenAI API & vLLM (OpenAI 兼容)
anthropic==0.18.1               # Anthropic Claude API
google-generativeai==0.3.2      # Google Gemini API

# Web 框架
flask==3.0.2                    # Web 框架
flask-cors==4.0.0               # CORS 支持
markdown==3.5.2                 # Markdown 渲染

# 定时任务
apscheduler==3.10.4             # 任务调度器

# 数据处理
pandas==2.2.1                   # 数据分析
requests==2.31.0                # HTTP 请求

# 工具库
tqdm==4.66.2                    # 进度条
colorama==0.4.6                 # 终端彩色输出

# 数据可视化和分析
wordcloud==1.9.3                # 词云生成
matplotlib==3.8.3               # 绘图库
pillow==10.2.0                  # 图像处理
jieba==0.42.1                   # 中文分词（用于中文词云）
scikit-learn==1.4.1.post1       # 机器学习（文本分析）
nltk==3.8.1                     # 自然语言处理
